template_hosting/python/requirements.txt
Flask
flask-cors
requests
pyyaml
rdflib
